Get rid of one level of abstraction
